New: Python 3 support.
Ran 2to3 on everything, corrected some incompatibilities by hand. Based on another project
that is using this dist and running on python3.5, seems to already be working very well.

Currently getting 10 failures and 41 errors on 160 tests. This is a great start. I'm sure a lot of these are incompatiblities as well!